requirement.txt
fastapi
uvicorn
sqlmodel
sqlalchemy
mysql-connector-python
Pillow
python-dotenv
python-jose
passlib
alembic